+++ src/dist_s1_enumerator/__init__.py
@@ -1,12 +1,15 @@
 import warnings
 from importlib.metadata import PackageNotFoundError, version
 
+import asf_search
+
 from dist_s1_enumerator.asf import (
     agg_rtc_metadata_by_burst_id,
     get_rtc_s1_metadata_from_acq_group,
     get_rtc_s1_ts_metadata_from_mgrs_tiles,
 )
 from dist_s1_enumerator.dist_enum import enumerate_dist_s1_products, enumerate_one_dist_s1_product
+from dist_s1_enumerator.dist_enum_inputs import enumerate_dist_s1_workflow_inputs
 from dist_s1_enumerator.mgrs_burst_data import (
     get_burst_ids_in_mgrs_tiles,
     get_burst_table,
@@ -30,14 +33,15 @@ except PackageNotFoundError:
         '   python -m pip install -e .\n',
         RuntimeWarning,
     )
-
+# Increase CMR timeout to 2 minutes
+asf_search.constants.INTERNAL.CMR_TIMEOUT = 120
 
 __all__ = [
     'agg_rtc_metadata_by_burst_id',
     'agg_rtc_metadata_by_burst_id',
     'enumerate_dist_s1_products',
     'enumerate_dist_s1_products',
-    'enumerate_one_dist_s1_product',
+    'enumerate_dist_s1_workflow_inputs',
     'enumerate_one_dist_s1_product',
     'get_burst_ids_in_mgrs_tiles',
     'get_burst_table_from_mgrs_tiles',